silo r_p_f, trouble testing in Rust
